+++ src/client/client.test.js
@@ -44,13 +44,41 @@ test('multiplayer server set when provided', () => {
   expect(client.multiplayerClient.socket.io.engine.port).toEqual(port);
 });
 
+test('accepts enhancer for store', () => {
+  let spyDispatcher;
+  const spyEnhancer = vanillaCreateStore => (...args) => {
+    const vanillaStore = vanillaCreateStore(...args);
+    return {
+      ...vanillaStore,
+      dispatch: (spyDispatcher = jest.fn(vanillaStore.dispatch)),
+    };
+  };
+  const client = Client({
+    game: Game({
+      moves: {
+        A: (G, ctx, arg) => ({ arg }),
+      },
+    }),
+    enhancer: spyEnhancer,
+  });
+
+  expect(spyDispatcher.mock.calls.length).toBe(0);
+  client.moves.A(42);
+  expect(spyDispatcher.mock.calls.length).toBe(1);
+});
+
 test('event dispatchers', () => {
   {
     const game = Game({});
     const reducer = createGameReducer({ game, numPlayers: 2 });
     const store = createStore(reducer);
     const api = createEventDispatchers(game.flow.eventNames, store);
-    expect(Object.getOwnPropertyNames(api)).toEqual(['endTurn']);
+    expect(Object.getOwnPropertyNames(api)).toEqual([
+      'undo',
+      'redo',
+      'endTurn',
+      'changeActionPlayers',
+    ]);
     expect(store.getState().ctx.turn).toBe(0);
     api.endTurn();
     expect(store.getState().ctx.turn).toBe(1);
@@ -59,8 +87,8 @@ test('event dispatchers', () => {
   {
     const game = Game({
       flow: {
-        undo: true,
         endPhase: true,
+        endGame: true,
       },
     });
     const reducer = createGameReducer({ game, numPlayers: 2 });
@@ -71,6 +99,8 @@ test('event dispatchers', () => {
       'redo',
       'endTurn',
       'endPhase',
+      'endGame',
+      'changeActionPlayers',
     ]);
     expect(store.getState().ctx.turn).toBe(0);
     api.endTurn();
@@ -82,7 +112,7 @@ test('event dispatchers', () => {
       flow: {
         endPhase: false,
         endTurn: false,
-        undo: false,
+        undoableMoves: [],
       },
 
       phases: [{ name: 'default' }],
@@ -90,7 +120,29 @@ test('event dispatchers', () => {
     const reducer = createGameReducer({ game, numPlayers: 2 });
     const store = createStore(reducer);
     const api = createEventDispatchers(game.flow.eventNames, store);
-    expect(Object.getOwnPropertyNames(api)).toEqual([]);
+    expect(Object.getOwnPropertyNames(api)).toEqual(['changeActionPlayers']);
+  }
+
+  {
+    const game = Game({
+      flow: {
+        endPhase: true,
+        undoableMoves: ['A'],
+      },
+    });
+    const reducer = createGameReducer({ game, numPlayers: 2 });
+    const store = createStore(reducer);
+    const api = createEventDispatchers(game.flow.eventNames, store);
+    expect(Object.getOwnPropertyNames(api)).toEqual([
+      'undo',
+      'redo',
+      'endTurn',
+      'endPhase',
+      'changeActionPlayers',
+    ]);
+    expect(store.getState().ctx.turn).toBe(0);
+    api.endTurn();
+    expect(store.getState().ctx.turn).toBe(1);
   }
 });