+++ src/core/flow.js
@@ -7,6 +7,38 @@
  */
 
 import { TurnOrder } from './turn-order';
+import { Random } from './random';
+import { Events } from './events';
+
+/**
+ * This function checks whether a player is allowed to make a move.
+ *
+ * @param {object} G     - The Game instance
+ * @param {object} ctx   - The ctx instance
+ * @param {object} opts  - Options - used here to transport the playerID.
+ */
+function canMakeMoveDefault(G, ctx, opts) {
+  const { playerID } = opts || {};
+
+  // In multiplayer mode, the default playerID is null, which corresponds
+  // to a spectator that can't make moves.
+  if (playerID === null) {
+    return true;
+  }
+
+  // In singleplayer mode (and most unit tests), the default playerID
+  // is undefined, and can always make moves.
+  if (playerID === undefined) {
+    return true;
+  }
+
+  const actionPlayers = ctx.actionPlayers || [];
+
+  // Explicitly do not allow the current player
+  // When he is allowed to make a move, his playerID
+  // must be included in actionPlayers.
+  return actionPlayers.includes(playerID) || actionPlayers.includes('any');
+}
 
 /**
  * Helper to create a reducer that manages ctx (with the
@@ -24,10 +56,6 @@ import { TurnOrder } from './turn-order';
  *                             reducer will handle. Each function
  *                             has the following signature:
  *                             ({G, ctx}) => {G, ctx}
- * @param {...object} validator - A move validator that returns false
- *                                if a particular type of move is invalid
- *                                at this point in the game.
- *                                (G, ctx, moveName) => boolean
  * @param {...object} processMove - A function that's called whenever a move is made.
  *                                  (state, action, dispatch) => state.
  * @param {...object} optimisticUpdate - (G, ctx, move) => boolean
@@ -36,20 +64,25 @@ import { TurnOrder } from './turn-order';
  *                                       the client while waiting for
  *                                       the result of execution from
  *                                       the server.
+ * @param {...object} canMakeMove - (G, ctx, opts) => boolean
+ *                                  Predicate to determine whether the player
+ *                                  identified by playerID is allowed to make a move.
+ *                                  opts contains an object { type, args, playerID },
+ *                                  which is the payload of makeMove().
  */
 export function Flow({
   ctx,
   events,
   init,
-  validator,
   processMove,
   optimisticUpdate,
+  canMakeMove,
 }) {
   if (!ctx) ctx = () => ({});
   if (!events) events = {};
   if (!init) init = state => state;
-  if (!validator) validator = () => true;
   if (!processMove) processMove = state => state;
+  if (!canMakeMove) canMakeMove = canMakeMoveDefault;
 
   if (optimisticUpdate === undefined) {
     optimisticUpdate = () => true;
@@ -71,13 +104,6 @@ export function Flow({
     ctx,
     init,
 
-    // Disallow moves once the game is over.
-    // Also call any provided additional validation.
-    validator: (G, ctx, move) => {
-      if (ctx.gameover !== undefined) return false;
-      return validator(G, ctx, move);
-    },
-
     eventNames: Object.getOwnPropertyNames(events),
 
     processMove: (state, action) => {
@@ -89,6 +115,13 @@ export function Flow({
     },
 
     optimisticUpdate,
+
+    // Disallow moves once the game is over.
+    // Also call any provided additional validation.
+    canMakeMove: (G, ctx, opts) => {
+      if (ctx.gameover !== undefined) return false;
+      return canMakeMove(G, ctx, opts);
+    },
   };
 }
 
@@ -130,7 +163,10 @@ export function Flow({
  *
  * @param {...object} endPhase - Set to false to disable the `endPhase` event.
  *
- * @param {...object} undo - Set to true to enable the undo/redo events.
+ * @param {...object} endGame - Set to true to enable the `endGame` event.
+ *
+ * @param {...object} undoableMoves - List of moves that are undoable,
+ *                                   (default: undefined, i.e. all moves are undoable).
  *
  * @param {...object} optimisticUpdate - (G, ctx, move) => boolean
  *                                       Control whether a move should
@@ -196,8 +232,10 @@ export function FlowWithPhases({
   turnOrder,
   endTurn,
   endPhase,
-  undo,
+  endGame,
+  undoableMoves,
   optimisticUpdate,
+  canMakeMove,
 }) {
   // Attach defaults.
   if (endPhase === undefined && phases) {
@@ -206,8 +244,8 @@ export function FlowWithPhases({
   if (endTurn === undefined) {
     endTurn = true;
   }
-  if (undo === undefined) {
-    undo = false;
+  if (endGame === undefined) {
+    endGame = false;
   }
   if (optimisticUpdate === undefined) {
     optimisticUpdate = () => true;
@@ -267,17 +305,28 @@ export function FlowWithPhases({
     return conf.endTurnIf(G, ctx);
   };
 
+  const getCurrentPlayer = (playOrder, playOrderPos) => {
+    if (playOrderPos === undefined) {
+      return 'any';
+    }
+    return playOrder[playOrderPos] + '';
+  };
+
   // Helper to perform start-of-phase initialization.
   const startPhase = function(state, phaseConfig) {
     const ctx = { ...state.ctx };
     const G = phaseConfig.onPhaseBegin(state.G, ctx);
-    ctx.currentPlayer = phaseConfig.turnOrder.first(G, ctx);
+    ctx.playOrderPos = phaseConfig.turnOrder.first(G, ctx);
+    ctx.currentPlayer = getCurrentPlayer(ctx.playOrder, ctx.playOrderPos);
+    ctx.actionPlayers = [ctx.currentPlayer];
     return { ...state, G, ctx };
   };
 
   const startTurn = function(state, config) {
-    const ctx = { ...state.ctx };
+    let ctx = { ...state.ctx };
     const G = config.onTurnBegin(state.G, ctx);
+    ctx = Random.detach(ctx);
+    ctx = Events.detach(ctx);
     const _undo = [{ G, ctx }];
     return { ...state, G, ctx, _undo, _redo: [] };
   };
@@ -365,11 +414,20 @@ export function FlowWithPhases({
     }
 
     // Update current player.
-    const currentPlayer = conf.turnOrder.next(G, ctx);
+    const playOrderPos = conf.turnOrder.next(G, ctx);
+    const currentPlayer = getCurrentPlayer(ctx.playOrder, playOrderPos);
+    const actionPlayers = [currentPlayer];
     // Update turn.
     const turn = ctx.turn + 1;
     // Update state.
-    ctx = { ...ctx, currentPlayer, turn, currentPlayerMoves: 0 };
+    ctx = {
+      ...ctx,
+      playOrderPos,
+      currentPlayer,
+      actionPlayers,
+      turn,
+      currentPlayerMoves: 0,
+    };
 
     // End phase if condition is met.
     const end = conf.endPhaseIf(G, ctx);
@@ -389,6 +447,11 @@ export function FlowWithPhases({
 
     const last = _undo[_undo.length - 1];
     const restore = _undo[_undo.length - 2];
+
+    // only allow undoableMoves to be undoable
+    if (undoableMoves && !undoableMoves.includes(last.moveType)) {
+      return state;
+    }
 
     return {
       ...state,
@@ -417,12 +480,22 @@ export function FlowWithPhases({
     };
   }
 
-  function processMove(state, action, dispatch) {
-    // Update currentPlayerMoves.
-    const currentPlayerMoves = state.ctx.currentPlayerMoves + 1;
-    state = { ...state, ctx: { ...state.ctx, currentPlayerMoves } };
+  function endGameEvent(state, arg) {
+    if (arg === undefined) {
+      arg = true;
+    }
 
+    return { ...state, ctx: { ...state.ctx, gameover: arg } };
+  }
+
+  function processMove(state, action, dispatch) {
     const conf = phaseMap[state.ctx.phase];
+
+    const currentPlayerMoves = state.ctx.currentPlayerMoves + 1;
+    state = {
+      ...state,
+      ctx: { ...state.ctx, currentPlayerMoves },
+    };
 
     const G = conf.onMove(state.G, state.ctx, action);
     state = { ...state, G };
@@ -435,14 +508,9 @@ export function FlowWithPhases({
       state = dispatch(state, { type: 'endTurn', playerID: action.playerID });
     }
 
-    // End the game automatically if endGameIf returns.
-    if (gameover !== undefined) {
-      return { ...state, ctx: { ...state.ctx, gameover } };
-    }
-
     // End the phase automatically if endPhaseIf is true.
     const end = conf.endPhaseIf(state.G, state.ctx);
-    if (end) {
+    if (end || gameover !== undefined) {
       state = dispatch(state, {
         type: 'endPhase',
         args: [end],
@@ -450,12 +518,18 @@ export function FlowWithPhases({
       });
     }
 
+    // End the game automatically if endGameIf returns.
+    if (gameover !== undefined) {
+      return { ...state, ctx: { ...state.ctx, gameover } };
+    }
+
     // Update undo / redo state.
     if (!endTurn) {
       const undo = state._undo || [];
+      const moveType = action.payload.type;
       state = {
         ...state,
-        _undo: [...undo, { G: state.G, ctx: state.ctx }],
+        _undo: [...undo, { G: state.G, ctx: state.ctx, moveType }],
         _redo: [],
       };
     }
@@ -463,22 +537,35 @@ export function FlowWithPhases({
     return state;
   }
 
-  const validator = (G, ctx, move) => {
-    const conf = phaseMap[ctx.phase];
+  const canMakeMoveWrap = (G, ctx, opts) => {
+    const conf = phaseMap[ctx.phase] || {};
     if (conf.allowedMoves) {
       const set = new Set(conf.allowedMoves);
-      return set.has(move.type);
+      if (!set.has(opts.type)) {
+        return false;
+      }
     }
-    return true;
+
+    // run user-provided validation
+    if (canMakeMove !== undefined && !canMakeMove(G, ctx, opts)) {
+      return false;
+    }
+
+    return canMakeMoveDefault(G, ctx, opts);
   };
 
   let enabledEvents = {};
-  if (undo) {
+  if (undoableMoves === undefined || undoableMoves.length > 0) {
     enabledEvents['undo'] = undoEvent;
     enabledEvents['redo'] = redoEvent;
   }
   if (endTurn) enabledEvents['endTurn'] = endTurnEvent;
   if (endPhase) enabledEvents['endPhase'] = endPhaseEvent;
+  if (endGame) enabledEvents['endGame'] = endGameEvent;
+
+  enabledEvents['changeActionPlayers'] = (state, actionPlayers) => {
+    return { G: state.G, ctx: { ...state.ctx, actionPlayers } };
+  };
 
   return Flow({
     ctx: numPlayers => ({
@@ -486,6 +573,8 @@ export function FlowWithPhases({
       turn: 0,
       currentPlayer: '0',
       currentPlayerMoves: 0,
+      playOrder: Array.from(Array(numPlayers), (d, i) => i),
+      playOrderPos: 0,
       phase: phases[0].name,
     }),
     init: state => {
@@ -499,7 +588,7 @@ export function FlowWithPhases({
       return optimisticUpdate(G, ctx, action);
     },
     events: enabledEvents,
-    validator,
     processMove,
+    canMakeMove: canMakeMoveWrap,
   });
 }